+++ client/src/App.js
@@ -1,14 +1,10 @@
 import React from "react";
 import "./Styles/Styles.css";
 import { Navbar } from "./components/Navbar/Navbar";
-// import { Search } from "./components/Search/Search";
-import { Exams } from "./components/Exams/Exams";
 import { Search } from "./components/Search/Search";
 import { Footer } from "./components/Footer/Footer";
-// import { CreateExamForm } from "./components/Create/CreateExamForm";
-// import { Edit } from "./components/Edit/Edit";
-// import { CreateExamForm } from "./components/Create/CreateExamForm";
-// import { NotFound } from "./components/Notfound/NotFound";
+import { CardData } from "./components/Cards/CardData";
+
 
 
 
@@ -18,7 +14,7 @@ const App = () => {
         <div className="App">
           <Navbar/>
           <Search/>
-          <Exams />
+          <CardData />
           <Footer/>
         </div>
       </>
