+++ client/src/components/Cards/CardData.js
@@ -1,107 +1,197 @@
-import React, { useState, useEffect } from 'react';
-import { Pagination } from '../Pagination/Pagination';
-import ExamsPopUp from '../PopUp/ExamsPopUp';
+import React, { useState, useEffect } from "react";
+import { Link } from "react-router-dom";
+import { AiOutlineEdit, AiOutlineClose } from "react-icons/ai"
+import { Pagination3 } from "../Pagination/Pagination3";
+
 
 export const CardData = () => {
-
-    {/*Using useState hook to initialize data state to an empty array.*/}
     const [data, setData] = useState([]);
+    const [selectedPatientId, setSelectedPatientId] = useState(null);
 
-    {/*Using useEffect hook to fetch data from an API and set the data state with the response data.*/}
+
+    // Fetching data from API using the useEffect Hook starts
     useEffect(() => {
-        fetch("https://czi-covid-lypkrzry4q-uc.a.run.app/api/exams")
+        fetch("http://localhost:9000/api/everything")
             .then((res) => res.json())
             .then((res) => setData(res.exams))
             .catch((error) => console.error("Error:", error));
     }, []);
+    // Fetching data from API using the useEffect Hook ends
 
-    {/*Using useState hook to initialize currentPage and itemsPerPage state to 1 and 16 respectively.*/}
-    const [currentPage, setCurrentPage] = useState(1);
-    const [itemsPerPage, setItemsPerPage] = useState(16);
+    // Function to group data by patient ID starts
+    const groupByPatientId = (data) => {
+        return data.reduce((acc, cur) => {
+            acc[cur.patientId] = acc[cur.patientId] || [];
+            acc[cur.patientId].push(cur);
+            return acc;
+        }, {});
+    };
+    const groupedData = groupByPatientId(data);
+    // Function to group data by patient ID ends
 
-    {/*Defining a function called handlePageChange to update currentPage state.*/}
-    const handlePageChange = (page) => {
-        setCurrentPage(page);
+    // Function to handle button click and set selected patient ID starts
+    const handleButtonClick = (patientId) => {
+        setSelectedPatientId(patientId);
     };
 
-    {/*Creating an empty array called pageNumbers. Looping through the data and calculating the total number of pages. Pushing each page number into the pageNumbers array.*/}
-    const pageNumbers = [];
-    for (let i = 1; i <= Math.ceil(data.length / itemsPerPage); i++) {
-        pageNumbers.push(i);
-    }
-
-    {/*Uses the reduce function to group the data array by patient ID. Initializes the accumulator object acc to an empty object. If the current patient ID is not in the accumulator object, initializes it with an object containing the patientId and an empty exams array. Pushes the current exam object to the corresponding patient object's exams array. Returns the accumulator object.*/}
-    const groupedData = data.reduce((acc, exam) => {
-        if (!acc[exam.patientId]) {
-            acc[exam.patientId] = { patientId: exam.patientId, exams: [] };
-        }
-        acc[exam.patientId].exams.push(exam);
-        return acc;
-    }, {});
-
-    {/*PopUp Window section.*/}
-    const [showMyData, setShowMyData] = useState(false)
-    const closeDataPopUp = () => setShowMyData(false);
-
-    {/*Defining a function called renderCard to render each patient card. Extracting patientId and exams from the patient object. Returning a button element with patient ID and styles.*/}
-    function renderCard(patient) {
-        const { patientId } = patient;
-        return (
-            <button key={patientId} onClick={() => setShowMyData(true)} className="card flex flex-col justify-center items-center border rounded-md shadow-md hover:shadow-lg hover:cursor-pointer space-y-1 h-16 md:h-20 lg:h-28 xl:h-32 2xl:space-y-6 2xl:h-72 ">
-                <th className="text-base font-bold lg:text-lg xl:text-2xl 2xl:text-5xl">Patient ID</th>
-                <td className="text-sm font-semibold lg:text-base xl:text-xl 2xl:text-4xl ">{patient.patientId}</td>
-            </button>
-        );
-    }
-
-    {/*Calculating the index of the last item on the current page.*/}
-    const indexOfLastItem = currentPage * itemsPerPage;
-    {/*Calculating the index of the first item on the current page.*/}
-    const indexOfFirstItem = indexOfLastItem - itemsPerPage;
-    {/*Slicing the grouped data based on the current page.*/}
-    const currentData = Object.values(groupedData).slice(indexOfFirstItem, indexOfLastItem);
-
-    {/*Defining a function called reduceRows to reduce the number of rows of patient cards. It takes two parameters, rowData and numColumns, which are used to calculate the number of rows and columns needed to display the data.*/}
-    const reduceRows = (rowData, numColumns) => {
-        const numRows = Math.ceil(rowData.length / numColumns);
-        const result = Array.from({ length: numRows }, (_, i) =>
-            rowData.slice(i * numColumns, i * numColumns + numColumns)
-        );
-        return result;
+    // Function to clear selected patient ID starts
+    const handleCloseModal = () => {
+        setSelectedPatientId(null);
     };
-
-    {/*Checking the device width to determine the number of columns and rows to display the data.*/}
-    const deviceWidth = window.innerWidth;
-    let numColumns, numRows;
-    if (deviceWidth >= 1024) {
-        // Desktop layout
-        numColumns = 4;
-        numRows = 8;
-    } else if (deviceWidth >= 768 && deviceWidth < 1024) {
-        // Tablet layout
-        numColumns = 2;
-        numRows = 6;
-    } else {
-        // Mobile layout
-        numColumns = 1;
-        numRows = 8;
-    }
-
-    {/*Reduce the data rows of data based on the number of columns specified. This function takes two arguments, an array of data and the number of columns to display, and returns a new array of data with a reduced number of rows.*/}
-    const reducedData = reduceRows(currentData, numColumns).slice(0, numRows);
 
     return (
-        <div>
-            <div className={`grid grid-cols-${numColumns} grid-rows-${numRows}  2xl:w-full mx-auto rounded-md p-4 h-auto gap-2`}>
-                {reducedData.map((row, index) => (
-                    <div key={index} className="grid grid-cols-1 gap-y-1">
-                        {row.map(renderCard)}
-                        <ExamsPopUp onClose={closeDataPopUp } visible={showMyData} />
-                    </div>
+        <>
+            <div className="mt-2 grid grid-cols-1 gap-4  p-3 sm:grid-cols-1 md:grid-cols-2 md:p-6 lg:grid-cols-3 lg:w-full lg:px-6 lg:gap-6 xl:grid-cols-4 2xl:px-20 2xl:gap-8">
+                {Object.entries(groupedData).map(([patientId, exams]) => (
+                    <button
+                        key={patientId}
+                        className=" flex items-center justify-center border-2 border-zinc-200 bg-zinc-100 rounded-lg shadow-xl hover:shadow-2xl focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2 transform transition hover:-translate-y-1 hover:scale-105 h-24  md:h-28 lg:h-36 2xl:h-60"
+                        onClick={() => handleButtonClick(patientId)}
+                    >
+
+                        <div className="flex flex-col items-center">
+                            <tr className="mb-2 lg:mb-1 xl:mb-2 2xl:mb-6">
+                                <thead className="">
+                                <th className="text-base font-bold text-gray-900 md:text-lg lg:text-xl xl:text-2xl 2xl:text-5xl">
+                                    Patient ID
+                                </th>
+                                </thead>
+                            </tr>
+                            <tr>
+                                <tbody>
+                                <td className="text-sm font-medium text-gray-90 md:text-sm lg:text-base xl:text-lg 2xl:text-4xl">
+                                    {patientId}
+                                </td>
+                                </tbody>
+                            </tr>
+                        </div>
+
+                    </button>
                 ))}
             </div>
-            {/* The Pagination component is used to display page numbers and allow the user to navigate through the data. It takes three props: currentPage which is the current page number, totalPages which is the total number of pages, and onPageChange which is a function that is called when the user clicks on a page number. This function is used to update the currentPage state variable in the parent component. */}
-            <Pagination pageNumbers={pageNumbers} currentPage={currentPage} onPageChange={handlePageChange} />
-        </div>
+            {/* Modal window */}
+            {selectedPatientId && (
+                <div className="fixed top-0 left-0 w-full h-full bg-gray-800 bg-opacity-30 backdrop-blur-sm flex justify-center items-center lg:px-16 xl:px-16 2xl:px-20">
+                    <div className=" bg-white rounded-xl shadow-md p-4 w-full h-1/2 overflow-y-auto">
+                        <div className="flex items-center justify-between  md:px-4 md:pt-4">
+                            <button
+                                className=" "
+                                onClick={() => setSelectedPatientId(null)}
+                            >
+                                <AiOutlineEdit className="text-gray-500 font-medium hover:text-gray-800 focus:outline-none text-base md:text-2xl xl:text-2xl 2xl:text-5xl" />
+                                {/* <EditExam patientId={selectedPatientId} onClose={() => setSelectedPatientId(null)} /> */}
+                            </button>
+                            <button
+                                className=""
+                                onClick={() => setSelectedPatientId(null)}
+                            >
+                                <AiOutlineClose className="text-gray-500 font-medium hover:text-gray-800 focus:outline-none text-base md:text-2xl xl:text-2xl 2xl:text-5xl" />
+                            </button>
+                        </div>
+
+                        <div className="flex justify-center items-center  px-10  2xl:h-1/5">
+                            <table className="flex flex-col justify-center items-center px-10  my-4 xl:gap-y-2 2xl:mb-10 2xl:gap-y-4 2xl:h-1/5">
+                                <tr className="flex flex-col items-center 2xl:mb-4 2xl:gap-4">
+                                    <thead>
+                                    <th className="text-lg font-semibold  text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                        Patient ID
+                                    </th>
+                                    </thead>
+                                </tr>
+                                <tbody className="">
+                                <tr className="">
+                                    <td className="w-full text-center text-sm font-semibold whitespace-normal break-words text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                        {selectedPatientId}
+                                    </td>
+                                </tr>
+                                </tbody>
+                            </table>
+                        </div>
+
+                        <div className="flex justify-center items-center 2xl:mt-4 2xl:h-1/2">
+                            {groupedData[selectedPatientId].map((exams) => (
+                                <div key={exams.id} className=" border-gray-200 h-full w-full">
+                                    <div className="w-full  mb-6 2xl:mb-16">
+                                        <table className="flex flex-col items-center xl:gap-y-2 2xl:gap-y-4 2xl:mb-6">
+                                            <thead className="w-full">
+                                            <tr className="grid grid-cols-6 text-center w-full mb-2">
+                                                <th className="text-sm  font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Exam ID
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Brixia Score
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Zip Code
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Sex
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Age
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    BMI
+                                                </th>
+                                            </tr>
+                                            </thead>
+                                            <tbody className="w-full">
+                                            <tr className=" grid grid-cols-6 text-center w-full whitespace-normal break-words">
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.examId}
+                                                </td>
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.brixiaScores}
+                                                </td>
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.zipCode}
+                                                </td>
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.sex}
+                                                </td>
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.age}
+                                                </td>
+                                                <td className="text-sm font-medium text-gray-500 md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl">
+                                                    {exams.bmi}
+                                                </td>
+                                            </tr>
+                                            </tbody>
+                                        </table>
+                                    </div>
+                                    <div className="w-full h-3/4 ">
+                                        <table className="flex flex-col  items-center xl:gap-y-2 2xl:gap-y-6">
+                                            <thead className="w-full">
+                                            <tr className="grid grid-cols-2 text-center w-full mb-2">
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Key Findings
+                                                </th>
+                                                <th className="text-sm font-medium text-gray-900 md:text-xl lg:text-xl xl:text-2xl 2xl:text-6xl">
+                                                    Images
+                                                </th>
+                                            </tr>
+                                            </thead>
+                                            <tbody className="w-full h-auto">
+                                            <tr className=" grid grid-cols-2 text-center w-full ">
+                                                <td className="text-sm font-medium text-gray-500 whitespace-normal break-words md:text-lg lg:text-lg xl:text-xl 2xl:text-5xl lg:mx-1.5 xl:mx-10 2xl:mx-10">
+                                                    {exams.keyFindings}
+                                                </td>
+                                                <td className="mx-auto">
+                                                    <Link className="flex justify-center border-2">
+                                                        <img src={exams.imageURL} alt="images" className="rounded-lg shadow-xl hover:shadow-2xl w-32 md:w-36 lg:w-40 xl:w-52 2xl:w-96 " />
+                                                    </Link>
+                                                </td>
+                                            </tr>
+                                            </tbody>
+                                        </table>
+                                    </div>
+                                </div>
+                            ))}
+                        </div>
+                    </div>
+                </div>
+            )}
+            <Pagination3/>
+        </>
     );
 };
